py-backend/requirements.txt
nltk
PyPDF2
flask
flask_cors
gunicorn==19.7.1
google-generativeai
python-dotenv

